pyproject.toml
[project]
name="mail-to-sqlite"
version="0.0.0"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4==4.12.2",
    "bs4==0.0.1",
    "cachetools==5.3.2",
    "certifi==2023.11.17",
    "charset-normalizer==3.3.2",
    "google-api-core==2.15.0",
    "google-api-python-client==2.112.0",
    "google-auth==2.26.1",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",
    "googleapis-common-protos==1.62.0",
    "httplib2==0.22.0",
    "idna==3.6",
    "oauthlib==3.2.2",
    "peewee==3.17.0",
    "protobuf==4.25.1",
    "pyasn1==0.5.1",
    "pyasn1-modules==0.3.0",
    "pyparsing==3.1.1",
    "requests==2.31.0",
    "requests-oauthlib==1.3.1",
    "rsa==4.9",
    "soupsieve==2.5",
    "uritemplate==4.1.1",
    "urllib3==2.1.0",
]

[project.scripts]
mail_to_sqlite = "mail_to_sqlite:main.main"

[tool.uv]
package = true
